requirements.txt
appdirs==1.4.0
packaging==16.8
PyGithub==1.29
pyparsing==2.1.10
six==1.10.0
